requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.0.0
python-multipart==0.0.12
PyJWT>=2.8.0
httpx>=0.27.0
